trash/requirements.txt
functions-framework==3.*
duckdb==1.1.1
google-cloud-secret-manager
feedparser
pandas
